Wait until STDOUT of the child process is closed before responding.
When the child process ends fast enough, the exit event of the child process is
emitted before all the data it put on STDOUT is read into the buffers.

This yields an incomplete response, which we do not want of course.
To reproduce: just run the tests on a machine that is fast enough.